chore: move PortableOpenApiErrorResponseAttributeBase to own file
